chore(erc-test-contracts): remove erc-test-contracts package
