--- packages/erc-test-contracts/.env.test
@@ -1,2 +1,0 @@
-WEB3="http://localhost:8552"
-DEPLOY_KEY="d9066ff9f753a1898709b568119055660a77d9aae4d7a4ad677b8fb3d2a571e5"
