--- packages/erc-test-contracts/.eslintignore
@@ -1,4 +1,0 @@
-node_modules
-build
-dist
-src/test
