--- packages/erc-test-contracts/.eslintrc.js
@@ -1,13 +1,0 @@
-module.exports = {
-    "extends": [
-        "../../.eslintrc.js"
-    ],
-    "parserOptions": {
-        "project": './tsconfig.build.json'
-    },
-    "rules": {
-        "@typescript-eslint/camelcase": "off",
-        "@typescript-eslint/no-explicit-any": "off",
-        "no-param-reassign": "off"
-    }
-};
