--- packages/erc-test-contracts/.lintstagedrc
@@ -1,6 +1,0 @@
-{
-  "*.{ts,tsx}": [
-    "eslint --fix",
-    "git add"
-  ]
-}
