--- packages/erc-test-contracts/.soliumignore
@@ -1,2 +1,0 @@
-node_modules
-contracts/Migrations.sol
