--- packages/erc-test-contracts/.soliumrc.json
@@ -1,21 +1,0 @@
-{
-  "extends": "solium:recommended",
-  "plugins": [
-    "security"
-  ],
-  "rules": {
-    "quotes": [
-      "error",
-      "double"
-    ],
-    "indentation": [
-      "error",
-      4
-    ],
-    "linebreak-style": [
-      "error",
-      "unix"
-    ],
-    "security/no-inline-assembly": "off"
-  }
-}
