--- packages/erc-test-contracts/CHANGELOG.md
@@ -1,40 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.2.2](https://github.com/energywebfoundation/origin/compare/@energyweb/erc-test-contracts@1.2.1...@energyweb/erc-test-contracts@1.2.2) (2019-11-11)
-
-
-### Bug Fixes
-
-* fix yarn clean command ([#224](https://github.com/energywebfoundation/origin/issues/224)) ([e07e9d8](https://github.com/energywebfoundation/origin/commit/e07e9d85de1b80c9f1a721398e41d82db580049c))
-
-
-
-
-
-## [1.2.1](https://github.com/energywebfoundation/origin/compare/@energyweb/erc-test-contracts@1.2.0...@energyweb/erc-test-contracts@1.2.1) (2019-10-25)
-
-
-### Bug Fixes
-
-* fix npm access on erc-test-token ([55814f5](https://github.com/energywebfoundation/origin/commit/55814f5aa78ca3b8d3c4cf91e3880761ded41f56))
-* use npm solc instead of downloading it every time ([#170](https://github.com/energywebfoundation/origin/issues/170)) ([e5ef574](https://github.com/energywebfoundation/origin/commit/e5ef574f6d297107606a1d035a56da01806a07d1))
-
-
-
-
-
-# 1.2.0 (2019-10-23)
-
-
-### Bug Fixes
-
-* **deps:** update dependency fs-extra to v8.1.0 ([#64](https://github.com/energywebfoundation/origin/issues/64)) ([ccd9ed5](https://github.com/energywebfoundation/origin/commit/ccd9ed5))
-
-
-### Features
-
-* demand fill tracking ([#82](https://github.com/energywebfoundation/origin/issues/82)) ([2948e17](https://github.com/energywebfoundation/origin/commit/2948e17))
-* erc-test-contracts ([#24](https://github.com/energywebfoundation/origin/issues/24)) ([ad771ab](https://github.com/energywebfoundation/origin/commit/ad771ab))
