--- packages/erc-test-contracts/README.md
@@ -1,6 +1,0 @@
-# Energy Web Foundation: ERC Test Contracts 
-
-## Warning
-
-This repository/package is not designed for productive usage. Instead it should be used for demonstration and learning purposes.
-Those contracts should only be used to test the ERC20 capability of the origin-release-B project.
