--- packages/erc-test-contracts/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@energyweb/erc-test-contracts",
-  "description": "",
-  "license": "GPL-3.0-only",
-  "author": "",
-  "homepage": "https://github.com/energywebfoundation/origin/tree/master/packages/erc-test-contracts#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/energywebfoundation/origin.git"
-  },
-  "bugs": {
-    "url": "https://github.com/energywebfoundation/origin/issues"
-  },
-  "version": "1.2.2",
-  "main": "dist/js/src/index.js",
-  "scripts": {
-    "build": "yarn build:ts",
-    "build:static": "yarn compile",
-    "build:ts": "tsc -b tsconfig.build.json --verbose --pretty",
-    "compile": "truffle compile",
-    "deploy-contracts": "truffle migrate",
-    "test": "mocha -r ts-node/register src/test/*.test.ts --timeout 90000 --exit",
-    "test:concurrent": "concurrently --success first --kill-others -n eth,test \"yarn start-ganache\" \"wait-on tcp:8552 && yarn test\"",
-    "start-ganache": "ganache-cli -q -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20 -p 8552",
-    "prettier": "prettier --write --config-precedence file-override './src/**/*'",
-    "lint": "solium -d contracts && eslint \"src/**/*{.ts,.tsx}\"",
-    "lint-fix": "solium -d contracts --fix && eslint --fix \"src/**/*{.ts,.tsx}\"",
-    "precommit": "lint-staged",
-    "clean": "shx rm -rf build dist"
-  },
-  "dependencies": {
-    "@energyweb/utils-general": "1.3.1",
-    "typedarray-to-buffer": "3.1.5",
-    "web3": "1.2.1",
-    "winston": "3.2.1",
-    "yaeti": "1.0.2"
-  },
-  "devDependencies": {
-    "@types/web3": "1.0.20",
-    "fs-extra": "8.1.0",
-    "ganache-cli": "6.4.5",
-    "solium": "1.2.5"
-  },
-  "postinstall": "rm -f node_modules/web3/index.d.ts",
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  }
-}
