--- packages/erc-test-contracts/tsconfig.build.json
@@ -1,14 +1,0 @@
-{
-    "extends": "../../tsconfig.build.json",
-    "compilerOptions": {
-        "composite": true,
-        "rootDir": "./",
-        "outDir": "./dist/js"
-    },
-    "include": ["./src/**/*", "./build/**/*.json"],
-    "references": [
-        {
-            "path": "../utils-general/tsconfig.build.json"
-        }
-    ]
-}
