--- packages/erc-test-contracts/tsconfig.json
@@ -1,5 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-    }
-}
