+++ packages/market/contracts.ts
@@ -1,4 +1,8 @@
 import MarketLogicJSON from './build/contracts/MarketLogic.json';
-import { migrateMarketRegistryContracts } from './src/utils/migrateContracts';
+import Erc20TestTokenJSON from './build/contracts/Erc20TestToken.json';
 
-export { MarketLogicJSON, migrateMarketRegistryContracts };
+export { MarketLogicJSON, Erc20TestTokenJSON };
+
+export { deployERC20TestToken } from './src/utils/deployERC20TestToken';
+export { Erc20TestToken } from './src/wrappedContracts/Erc20TestToken';
+export { migrateMarketRegistryContracts } from './src/utils/migrateContracts';
