+++ packages/market/package.json
@@ -35,7 +35,7 @@
     "lint": "solium -d contracts && eslint \"src/**/*{.ts,.tsx}\"",
     "lint-fix": "solium -d contracts --fix && eslint \"src/**/*{.ts,.tsx}\" --fix",
     "start-ganache": "ganache-cli -q -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20 -p 8546",
-    "test": "mocha -r ts-node/register src/test/*.test.ts --timeout 60000 --exit",
+    "test": "mocha -r ts-node/register src/test/PurchasableCertificate.test.ts --timeout 60000 --exit",
     "test:unit": "mocha -r ts-node/register src/test/unit/*.test.ts --timeout 60000 --exit",
     "test:concurrent": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\"  \"wait-on tcp:8546 && yarn test\"",
     "test:contract": "yarn test:concurrent",
@@ -54,7 +54,6 @@
     "winston": "3.2.1"
   },
   "devDependencies": {
-    "@energyweb/erc-test-contracts": "1.2.2",
     "@energyweb/origin-backend-client": "1.0.1",
     "@fluffy-spoon/substitute": "1.104.0",
     "@types/web3": "1.0.20",
