+++ packages/market/src/test/PurchasableCertificate.test.ts
@@ -8,11 +8,12 @@ import { migrateUserRegistryContracts } from '@energyweb/user-registry/contracts
 import { Asset, ProducingAsset, AssetLogic } from '@energyweb/asset-registry';
 import { migrateAssetRegistryContracts } from '@energyweb/asset-registry/contracts';
 import { Configuration, Compliance, Currency } from '@energyweb/utils-general';
-import { deployERC20TestToken, Erc20TestToken } from '@energyweb/erc-test-contracts';
 import { Certificate, CertificateLogic } from '@energyweb/origin';
 import { migrateCertificateRegistryContracts } from '@energyweb/origin/contracts';
 import { OffChainDataClientMock } from '@energyweb/origin-backend-client';
 
+import { deployERC20TestToken } from '../utils/deployERC20TestToken';
+import { Erc20TestToken } from '../wrappedContracts/Erc20TestToken';
 import { PurchasableCertificate, MarketLogic } from '..';
 import { migrateMarketRegistryContracts } from '../utils/migrateContracts';
 import { logger } from '../Logger';
