+++ packages/market/src/utils/deployERC20TestToken.ts
@@ -1,10 +1,9 @@
-import { deploy } from '@energyweb/utils-general';
-// eslint-disable-next-line import/no-unresolved
-import { TransactionReceipt } from 'web3/types';
-
 import Web3 from 'web3';
-import { Erc20TestTokenJSON, Erc721TestReceiverJSON } from '..';
+
+import { deploy } from '@energyweb/utils-general';
 import { Erc20TestToken } from '../wrappedContracts/Erc20TestToken';
+
+import Erc20TestTokenJSON from '../../build/contracts/Erc20TestToken.json';
 
 export async function deployERC20TestToken(
     web3: Web3,
@@ -26,19 +25,4 @@ export async function deployERC20TestToken(
     await erc20.initialize(testAccount);
 
     return contractAddress;
-}
-
-export async function deployERC721TestReceiver(
-    web3: Web3,
-    erc721TokenContract: string,
-    deployKey: string
-): Promise<TransactionReceipt> {
-    const privateKeyDeployment = deployKey.startsWith('0x') ? deployKey : `0x${deployKey}`;
-
-    return deploy(
-        web3,
-        Erc721TestReceiverJSON.bytecode +
-            web3.eth.abi.encodeParameter('address', erc721TokenContract).substr(2),
-        { privateKey: privateKeyDeployment }
-    );
 }
