+++ packages/origin-ui/package.json
@@ -25,7 +25,6 @@
   "dependencies": {
     "@date-io/moment": "1.3.11",
     "@energyweb/asset-registry": "1.3.3",
-    "@energyweb/erc-test-contracts": "1.2.2",
     "@energyweb/market": "1.3.3",
     "@energyweb/market-matcher": "1.3.3",
     "@energyweb/origin": "1.3.3",
