+++ packages/origin/package.json
@@ -46,7 +46,6 @@
     "winston": "3.2.1"
   },
   "devDependencies": {
-    "@energyweb/erc-test-contracts": "1.2.2",
     "@energyweb/origin-backend-client": "1.0.1",
     "@types/web3": "1.0.20",
     "ethlint": "1.2.5",
