+++ packages/origin/tsconfig.build.json
@@ -8,9 +8,6 @@
     "include": ["./contracts.ts", "./src/**/*", "./dist/ts/**/*", "./build/**/*.json"],
     "references": [
         {
-            "path": "../erc-test-contracts/tsconfig.build.json"
-        },
-        {
             "path": "../user-registry/tsconfig.build.json"
         },
         {
