+++ packages/origin/tsconfig.build.shakeable.json
@@ -6,9 +6,6 @@
     },
     "references": [
         {
-            "path": "../erc-test-contracts/tsconfig.build.json"
-        },
-        {
             "path": "../user-registry/tsconfig.build.shakeable.json"
         },
         {
