+++ packages/utils-demo/src/market.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import Web3 from 'web3';
 
-import { deployERC20TestToken, Erc20TestToken } from '@energyweb/erc-test-contracts';
+import { Erc20TestToken, deployERC20TestToken } from '@energyweb/market/contracts';
 import { Configuration, TimeFrame, Compliance, Currency } from '@energyweb/utils-general';
 import { User, UserLogic, Role, buildRights } from '@energyweb/user-registry';
 import { AssetLogic } from '@energyweb/asset-registry';
