+++ packages/utils-demo/tsconfig.build.json
@@ -7,9 +7,6 @@
     "include": ["src/**/*"],
     "references": [
         {
-            "path": "../erc-test-contracts/tsconfig.build.json"
-        },
-        {
             "path": "../user-registry/tsconfig.build.json"
         },
         {
